HUE-2761 [notebook] Snippet resize can get stuck
When pasting content that resizes a snippet editor the element that is used to resize the editor can get out of position. The jQuery draggable plugin sets an inline style on the element once it's dragged, giving it a fixed top value which put the element out of position when the editor height changed. To solve this the inline style is cleared on drag stop.

This commit also fixes a related issue with an incorrect start position of the draggable element on drag start. This was caused by the codemirrorSize observable being out of sync with the editor height after paste, the codemirrorSize is used to determine the drag start position on the draggable element.